+++ packages/wagtail/.storybook/__mocks__/eventEmitterMock.ts
@@ -5,18 +5,19 @@ export class EventEmitter {
     if (!this.events[event]) {
       this.events[event] = []
     }
-    this.events[event].push(listener)
+    this.events[event]?.push(listener)
   }
 
   emit(event: string, ...args: any[]) {
     if (this.events[event]) {
-      this.events[event].forEach((listener) => listener(...args))
+      this.events[event]?.forEach((listener) => listener(...args))
     }
   }
 
   off(event: string, listener: Function) {
     if (this.events[event]) {
-      this.events[event] = this.events[event].filter((l) => l !== listener)
+      // @ts-ignore
+      this.events[event] = this.events[event]?.filter((l) => l !== listener)
     }
   }
 }