+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @baseapp-frontend/utils
 
+## 3.0.5
+
+### Patch Changes
+
+- Remove events from catalogs.
+- Use stable versions on core dependencies.
+
 ## 3.0.4
 
 ### Patch Changes